+++ noodles-cram/src/async/io/reader/header/container/sam_header.rs
@@ -27,6 +27,24 @@ where
             is_eol: true,
         }
     }
+
+    pub async fn discard_to_end(&mut self) -> io::Result<usize> {
+        let mut n = 0;
+
+        loop {
+            let src = self.inner.fill_buf().await?;
+
+            if src.is_empty() {
+                return Ok(n);
+            }
+
+            let len = src.len();
+
+            self.inner.consume(len);
+
+            n += len;
+        }
+    }
 }
 
 impl<R> AsyncRead for Reader<R>