+++ noodles-cram/src/io/reader/header.rs
@@ -88,6 +88,7 @@ where
 
     let mut raw_sam_header_reader = container_reader.raw_sam_header_reader()?;
     let header = read_sam_header(&mut raw_sam_header_reader)?;
+    raw_sam_header_reader.discard_to_end()?;
 
     Ok(header)
 }